+++ include/mergebot/core/model/node/LinkageSpecNode.h
@@ -15,7 +15,7 @@ public:
                   const std::string &QualifiedName,
                   const std::string &OriginalSignature, std::string &&Comment,
                   const std::optional<ts::Point> &Point, std::string &&USR,
-                  size_t BeforeFirstChildEOLs, size_t ParentSignatureHash,
+                  int BeforeFirstChildEOLs, size_t ParentSignatureHash,
                   bool IsSynthetic = false)
       : CompositeNode(NodeId, NeedToMerge, Kind, DisplayName, QualifiedName,
                       OriginalSignature, std::move(Comment), Point,
